Return int64_t from SourceRegionContainer::index()
index(sr, g) computes sr * negroups_ + g in 64-bit (sr is int64_t) but truncated the result back to int on return. With more than ~30.7M source regions at 70 groups the flattened element index exceeds 2^31, yielding a wrapped/negative offset and an out-of-bounds access into the correctly-sized flux arrays during the solve. Return int64_t.

+++ include/openmc/random_ray/source_region.h
@@ -690,7 +690,7 @@ private:
   // Private Methods
 
   // Helper function for indexing
-  inline int index(int64_t sr, int g) const { return sr * negroups_ + g; }
+  inline int64_t index(int64_t sr, int g) const { return sr * negroups_ + g; }
 };
 
 } // namespace openmc
